Related work items: #59158, #60800, #60803, #60904, #60909, #60910, #60911, #60912, #60990, #60991, #61012

+++ Logitude/Logitude.Accounting.Data/EntityPOCOs/GLAccount.cs
@@ -179,7 +179,7 @@ namespace Logitude.Accounting.Data.EntityPOCOs
         [Column("ActiveForInterestCreditInvoice")]
 	    public bool? ActiveForInterestCreditInvoice { get; set; }
         [Column("MinimumInterestInvoiceBilling")]
-	    public decimal? MinimumInterestInvoiceBilling { get; set; }
+	    public int? MinimumInterestInvoiceBilling { get; set; }
         [Column("InterestCreditLimit")]
 	    public decimal? InterestCreditLimit { get; set; }
     }
